Adding Webworks 1.x and 2.x support for BB10
- Cordova RTC work item number: 59037
- Apache Cordova Jira issue:
https://issues.apache.org/jira/browse/CB-5916
- Markmail / Google Group / StackOverflow reference:
- Product PMR: 41288,057,649
- Worklight RTC work item number: 58719
- Description: After fix the window.blackberry.invoke(), part of the
blackberry plugins,  an optimization to the added code was required to
3.4.0esr, and it's required by 3.6.1esr as well.
window.webworks.exec is non-required, what's required is 'webworksready'
event.
As part of the optimization, support for Webworks 2.x was added in the
code.
Now projects can handle WW 1.x and in case of upgrade, it can support WW
2.x as well.